Version 2.01 : BuildKite and Replication integrations control
* Allow disabling BuildKite integration

Allows the replication or buildkite integration feature to be selectively enabled with `--enable-change-replication` and `--enable-buildkite-integration` respectively.

It might be nice to convert `--disable-buildkite-webhook-handler` to use the same `--enable-...` idiom.

* Per Change replication

This lets `patchset-created` events trigger a replication to a remote repository.

* fix: Source and Destination credentials

SSH replication should use different credentials for the source and for the destination.

* Add test environment

Adds the Docker Compose for standing up a testing environment.
Adds the `ssh-server` container resources for building a local git SSH service.

* Add build information and basic usage

* Gerrit service for development

Docker Compose can run a `gerrit` service using configuration from `.env` with `docker-compose up -d gerrit`.

The Gerrit instance and state can be recreated at any time. Running `setup_gerrit.sh` creates an SSH key pair and authorizes it as the Gerrit admin to the instance and then creates a test project with an initial remote and local commit. The `pre_merge` workflow has a full demo.